Feature #70119, refactor VSInputService as proxy and update related controllers
