+++ core/src/main/java/inetsoft/web/binding/VSBindingController.java
@@ -17,13 +17,15 @@
  */
 package inetsoft.web.binding;
 
+import inetsoft.util.Tool;
 import inetsoft.web.binding.model.*;
-import inetsoft.web.binding.service.VSBindingService;
+import inetsoft.web.binding.service.VSBindingServiceProxy;
 import inetsoft.web.composer.model.vs.VSTableTrapModel;
 import inetsoft.web.factory.RemainingPath;
 import inetsoft.web.viewsheet.LoadingMask;
 import inetsoft.web.viewsheet.Undoable;
 import inetsoft.web.viewsheet.event.InsertSelectionChildEvent;
+import inetsoft.web.viewsheet.model.RuntimeViewsheetRef;
 import inetsoft.web.viewsheet.service.CommandDispatcher;
 import inetsoft.web.viewsheet.service.LinkUri;
 import org.slf4j.Logger;
@@ -40,10 +42,10 @@ import java.security.Principal;
 @RestController
 public class VSBindingController {
    @Autowired
-   public VSBindingController(VSBindingService vsBindingService,
-                              VSBindingControllerServiceProxy vsBindingServiceProxy)
+   public VSBindingController(VSBindingServiceProxy vsBindingServiceProxy,
+                              RuntimeViewsheetRef runtimeViewsheetRef)
    {
-      this.vsBindingService = vsBindingService;
+      this.runtimeViewsheetRef = runtimeViewsheetRef;
       this.vsBindingServiceProxy = vsBindingServiceProxy;
    }
 
@@ -100,6 +102,7 @@ public class VSBindingController {
                         Principal principal)
       throws Exception
    {
+      vsId = Tool.byteDecode(vsId);
       return vsBindingServiceProxy.commit(vsId, assemblyName, editMode, originalMode, principal);
    }
 
@@ -120,7 +123,8 @@ public class VSBindingController {
                            Principal principal, CommandDispatcher dispatcher)
       throws Exception
    {
-      this.vsBindingService.insertChild(event, linkUri, principal, dispatcher);
+      this.vsBindingServiceProxy.insertChild(runtimeViewsheetRef.getRuntimeId(), event,
+                                        linkUri, principal, dispatcher, false);
    }
 
    /**
@@ -140,10 +144,10 @@ public class VSBindingController {
                                                 Principal principal)
       throws Exception
    {
-      return vsBindingServiceProxy.checkVSSelectionTrap(runtimeId, event, linkUri, principal);
+      return vsBindingServiceProxy.checkVSSelectionTrap(runtimeId, event, principal);
    }
 
-   private final VSBindingService vsBindingService;
-   private final VSBindingControllerServiceProxy vsBindingServiceProxy;
+   private final VSBindingServiceProxy vsBindingServiceProxy;
+   private final RuntimeViewsheetRef runtimeViewsheetRef;
    private static final Logger LOG = LoggerFactory.getLogger(VSBindingController.class);
 }
