+++ core/src/main/java/inetsoft/web/composer/model/vs/VSTableTrapModel.java
@@ -18,9 +18,11 @@
 package inetsoft.web.composer.model.vs;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.immutables.serial.Serial;
 import org.immutables.value.Value;
 
 @Value.Immutable
+@Serial.Structural
 @JsonSerialize(as = ImmutableVSTableTrapModel.class)
 public abstract class VSTableTrapModel {
    public abstract boolean showTrap();
